updated addLyrics.py for syllables
The original addLyrics did not use the correct syllabic element, since there were not syllable markers retained for input.
This version uses text that has hyphen+space to separate syllables. Then the output uses "begin", "middle" and "end" syllabic text correctly, while removing the hyphens in the final xml, so that sinsy is more likely to pronounce the words correctly.
This version will also put in an "extend" element for syllables that end with an underscore, but currently there is no way of indicating how many notes the syllable should be extended to.